requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
certifi==2025.10.5
discord.py==2.6.4
dnspython==2.8.0
frozenlist==1.8.0
idna==3.11
motor==3.7.1
multidict==6.7.0
propcache==0.4.1
pymongo==4.15.3
python-dotenv==1.2.1
yarl==1.22.0
